Update Generator to place assets in raty/applicaiton subfolders
The original file setup pushed all of the Gem's assets into the
application's main asset folder. By moving these into subfolders,
there is less chance that a user will lose track of the assets,
overwrite their own assets or have the assets interfere with the
asset organization for the site.