Revert jquery.raty.js defaults to original state and move gem specific defaults into ratyrate.js.erb
The original jquery.raty plugin/script has a defaults object that was being overwritten by the ratyrate dev team. One of the settings was adding asset-path within this file. However, this file is not an ERB file so Rails was failing when it hit it. By pulling that code from the plugin and instead placing it in the ratyrate.js.erb file we are abstracting it out of the plugin. This fixes the issue that Rails has AND it also means that if the raty.js script is updated we could easily drop in the new version without worrying about any the defaults we needed for the ratyrate Gem.